requirements.txt
# ScholarMind Dependencies
# Install: pip install -r requirements.txt

# Core: OpenAI SDK (MiniMax 兼容) + MCP
openai>=1.0.0
mcp[cli]>=1.0.0

# PDF Parsing
PyMuPDF>=1.24.0

# Vector Database (RAG)
chromadb>=0.5.0

# Knowledge Graph
networkx>=3.2

# HTTP Client
httpx>=0.27.0

# Environment Variables
python-dotenv>=1.0.0

# Terminal UI
rich>=13.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0